/throw exception if null

+++ src/main/java/com/luv2code/springdemo/rest/CustomerRestController.java
@@ -71,6 +71,12 @@ public class CustomerRestController {
 	public String deleteCustomer(@PathVariable int customerId) {
 		
 		Customer tempCustomer = customerService.getCustomer(customerId);
+		
+		//throw exception if null
+		if(tempCustomer == null) {
+			throw new CustomerNotFoundException("Customer id not found - " + customerId);
+		}
+		
 		customerService.deleteCustomer(customerId);
 		
 		return "Deleted customer id - " + customerId;
